updating MongoDB

+++ DietManager/src/main/java/it/unipi/dii/dietmanager/persistence/MongoDB.java
@@ -1,6 +1,8 @@
 package it.unipi.dii.dietmanager.persistence;
 
 import com.mongodb.ConnectionString;
+import com.mongodb.ReadConcern;
+import com.mongodb.WriteConcern;
 import com.mongodb.bulk.BulkWriteResult;
 import com.mongodb.client.*;
 import com.mongodb.client.model.*;
@@ -13,8 +15,10 @@ import it.unipi.dii.dietmanager.entities.*;
 
 import java.util.*;
 
+import static com.mongodb.client.model.Aggregates.*;
 import static com.mongodb.client.model.Filters.and;
 import static com.mongodb.client.model.Filters.eq;
+import static com.mongodb.client.model.Projections.*;
 
 public class MongoDB{
 
@@ -38,6 +42,10 @@ public class MongoDB{
         ConnectionString connectionString = new ConnectionString("mongodb://localhost:" + localhostPort);
         mongoClient = MongoClients.create(connectionString);
         database = mongoClient.getDatabase("dietManagerDB");
+        database.withWriteConcern(WriteConcern.ACKNOWLEDGED);   // Write Concern to wait for acknowledgement according to Server configuration.
+                                                                // Needed for wasAcknowledged() methods.
+                                                                // other options are possible.
+        database.withReadConcern(ReadConcern.DEFAULT);          // Use the servers default read concern.
         return true;
     }
 
@@ -78,7 +86,7 @@ public class MongoDB{
     }
 
     // for both signIn and lookUpUserbyUsername
-    public User lookUpUserByID(String username){
+    public User lookUpUserByUsername(String username){
         openConnection();
         MongoCollection<Document> usersCollection = database.getCollection(COLLECTION_USERS);
         Document userDocument = usersCollection.find(eq(User.USERNAME, new ObjectId(username))).first(); // there can be at least only 1 match.
@@ -96,10 +104,19 @@ public class MongoDB{
 
     public boolean removeUser(String username){
         openConnection();
+        User userToRemove = lookUpUserByUsername(username);
+        if(userToRemove == null)
+            return false;
         MongoCollection<Document> usersCollection = database.getCollection(COLLECTION_USERS);
         DeleteResult deleteResult = usersCollection.deleteOne(eq(User.USERNAME, username));
+        boolean isSuccessful = deleteResult.wasAcknowledged();
+
+        // if user removed was a Nutritionist, all his diets are deleted altogether.
+        if(isSuccessful && userToRemove instanceof Nutritionist){
+            isSuccessful = removeDietsByNutritionist(username);
+        }
         closeConnection();
-        return deleteResult.wasAcknowledged();
+        return isSuccessful;
     }
 
     public List<Nutritionist> lookUpNutritionistsByCountry(String country){
@@ -143,20 +160,25 @@ public class MongoDB{
         return Document.parse(diet.toJSONObject().toString());
     }
 
+    // for both lookUpDietByID and lookUpStandardUserCurrentDiet
     public Diet lookUpDietByID(String id){
         openConnection();
-        MongoCollection<Document> dietCollection = database.getCollection(COLLECTION_DIETS);
-        Document dietDocument = dietCollection.find(eq(Diet.ID, new ObjectId(id))).first(); // there can be at least only 1 match.
+        Document dietDocument = null;
+        try {
+            MongoCollection<Document> dietCollection = database.getCollection(COLLECTION_DIETS);
+            dietDocument = dietCollection.find(eq(Diet.ID, new ObjectId(id))).first(); // there can be at least only 1 match.
+        }catch(Exception e){}
         closeConnection();
         return dietFromDocument(dietDocument);
     }
 
-    public List<Diet> lookUpDietByName(String name){
+    public List<Diet> lookUpDietByName(String subname){    // RETURN ALL DIETS whose name include substring seached.
         openConnection();
+        String regex = "\\.\\*"+subname+"\\.\\*";
         List<Diet> diets = new ArrayList<>();
         MongoCollection<Document> dietCollection = database.getCollection(COLLECTION_DIETS);
         try(MongoCursor<Document> cursor = dietCollection.find(
-                eq(Diet.NAME, name)).iterator()){
+                eq(Diet.NAME, regex)).iterator()){
             while(cursor.hasNext()){
                 diets.add(dietFromDocument(cursor.next()));
             }
@@ -181,7 +203,6 @@ public class MongoDB{
 
     public boolean followDiet(StandardUser standardUser, String dietID){
         openConnection();
-        // verifica esistenza Dieta
         MongoCollection<Document> userCollection = database.getCollection(COLLECTION_USERS);
         Bson userFilter = Filters.eq( User.USERNAME, standardUser.getUsername() );
         Bson insertCurrentDietField = Updates.set(StandardUser.CURRENT_DIET, dietID);
@@ -205,8 +226,8 @@ public class MongoDB{
                 userFilter, deleteEatenFoodsField);
 
         List<WriteModel<Document>> bulkOperations = new ArrayList<>();
+        bulkOperations.addAll(Arrays.asList(updateRemoveCurrentDietDocument,updateRemoveEatenFoodsDocument));
         BulkWriteResult bulkWriteResult = userCollection.bulkWrite(bulkOperations);
-
         closeConnection();
         return bulkWriteResult.wasAcknowledged();
     }
@@ -214,46 +235,101 @@ public class MongoDB{
     public boolean addDiet(Diet diet){
         openConnection();
         MongoCollection<Document> dietCollection = database.getCollection(COLLECTION_DIETS);
-        InsertOneResult insertOneResult = dietCollection.insertOne(dietToDocument(diet));
+        Document dietDocument = dietToDocument(diet);
+        InsertOneResult insertOneResult = dietCollection.insertOne(dietDocument);
+        // adding '_id' to diet object passed as argument
+        diet.setId(dietDocument.getObjectId("_id").toString());
         closeConnection();
-        return insertOneResult.wasAcknowledged(); // before: lookUpDietByID(diet.getId()) != null
+        return insertOneResult.wasAcknowledged();
     }
 
     public boolean removeDiet(String dietID){
         openConnection();
+        // deleting 'diet' document from 'diets' collection
         MongoCollection<Document> dietCollection = database.getCollection(COLLECTION_DIETS);
-        DeleteResult deleteResult = dietCollection.deleteOne(eq(Diet.ID, dietID));
+        DeleteResult deleteDietResult = dietCollection.deleteOne(eq(Diet.ID, dietID));
+        boolean isSuccessful = deleteDietResult.wasAcknowledged();
+
+        // deleting 'currentDiet' field from each 'user' document which has as 'currentDiet' value the target diet to remove
+        if(isSuccessful){
+            MongoCollection<Document> userCollection = database.getCollection(COLLECTION_USERS);
+
+            Bson userFilter = Filters.eq( StandardUser.CURRENT_DIET, dietID );
+            Bson deleteCurrentDietField = Updates.unset(StandardUser.CURRENT_DIET);
+            Bson deleteEatenFoodsField = Updates.unset(StandardUser.EATENFOODS);
+
+            UpdateManyModel<Document> updateRemoveCurrentDietDocument = new UpdateManyModel<>(
+                    userFilter, deleteCurrentDietField);
+            UpdateManyModel<Document> updateRemoveEatenFoodsDocument = new UpdateManyModel<>(
+                    userFilter, deleteEatenFoodsField);
+
+            List<WriteModel<Document>> bulkOperations = new ArrayList<>();
+            bulkOperations.addAll(Arrays.asList(updateRemoveCurrentDietDocument,updateRemoveEatenFoodsDocument));
+            BulkWriteResult bulkWriteResult = userCollection.bulkWrite(bulkOperations);
+            isSuccessful = bulkWriteResult.wasAcknowledged();
+        }
         closeConnection();
-        return deleteResult.wasAcknowledged();
+        return isSuccessful;
     }
 
-    /*
+    private boolean removeDietsByNutritionist(String username){
+        List<Diet> dietsToRemove = lookUpDietByNutritionist(username);
+        boolean isSuccessful = true, isCurrentSuccessful;
+        for(Diet diet: dietsToRemove){
+            isCurrentSuccessful = removeDiet(diet.getId());
+            if(isCurrentSuccessful == false)
+                isSuccessful = false;
+        }
+        return isSuccessful;
+    }
+
+
     public HashMap<String, Nutrient> lookUpMostSuggestedNutrientForEachNutritionist(){
         openConnection();
+
+        // retrieve documents representing for each combination of Nutritionist-Nutrient, the average quantity of the nutrient
+        // that the nutritionist has recommended in its diets.
+        MongoCollection<Document> dietCollection = database.getCollection(COLLECTION_DIETS);
         HashMap<String, Nutrient> nutritionistNutrientMap = new HashMap<>();
 
         Bson dietsUnwindNutrients = unwind(Diet.NUTRIENTS);
-        Bson projection1 = project(fields(include(Diet.NUTRITIONIST,Nutrient.QUANTITY),computed("nutrientName",Diet.NUTRIENTS)));
-        Bson groupByNutrientAndNutritionist = new Document("$group",
-                new Document(Diet.NUTRITIONIST, "$"+Diet.NUTRITIONIST).append(Diet.NUTRIENTS,"$"+Diet.NUTRIENTS)
-                .append("totalQuantity", new Document("$sum","$"+Nutrient.QUANTITY)));
+        Bson projectionFields = project(
+                fields(include(Diet.NUTRITIONIST,Nutrient.QUANTITY, Nutrient.UNIT),computed("nutrientName",Diet.NUTRIENTS)));
+        Bson dietsMatchOutEnergyNutrient = match(
+                Filters.eq("nutrientName", "Energy"));
+        Bson groupByNutrientAndNutritionist = group(
+                new Document(Diet.NUTRITIONIST, "$"+Diet.NUTRITIONIST).append(Diet.NUTRIENTS,"$"+Diet.NUTRIENTS).append(Nutrient.UNIT,"$"+Nutrient.UNIT),
+                Accumulators.avg("averageQuantity", "$"+Nutrient.QUANTITY));
+        closeConnection();
 
-
-
-        List<Diet> diets = new ArrayList<>();
-        MongoCollection<Document> dietCollection = database.getCollection(COLLECTION_DIETS);
-        try(MongoCursor<Document> cursor = dietCollection.find(
-                eq(Diet.NUTRITIONIST, username)).iterator()){
+        // analyze for each nutritionist which nutrient he has recommended the most.
+        Document currentDocument;
+        List<String> nutritionists = new ArrayList<>();
+        List<String> nutrients = new ArrayList<>();
+        List<String> units = new ArrayList<>();
+        List<Double> quantities = new ArrayList<>();
+        try(MongoCursor<Document> cursor = dietCollection.aggregate(Arrays.asList(dietsUnwindNutrients,
+                                                projectionFields,
+                                                dietsMatchOutEnergyNutrient,
+                                                groupByNutrientAndNutritionist)).iterator()){
             while(cursor.hasNext()){
-                diets.add(dietFromDocument(cursor.next()));
+                currentDocument = cursor.next();
+                nutritionists.add(currentDocument.getString(Diet.NUTRITIONIST));
+                nutrients.add(currentDocument.getString("nutrientName"));
+                units.add(currentDocument.getString(Nutrient.UNIT));
+                quantities.add(currentDocument.getDouble(Nutrient.QUANTITY));
             }
         }
-        closeConnection();
+
         return nutritionistNutrientMap;
 
 
     }
-    */
+
+
+
+
+
 
     /************************************************************************************/
     /*************************** Foods related methods **********************************/
@@ -286,8 +362,8 @@ public class MongoDB{
         String regex = "\\.\\*"+name+"\\.\\*";
 
         List<Food> foods = new ArrayList<>();
-        MongoCollection<Document> dietCollection = database.getCollection(COLLECTION_FOODS);
-        try(MongoCursor<Document> cursor = dietCollection.find(
+        MongoCollection<Document> foodCollection = database.getCollection(COLLECTION_FOODS);
+        try(MongoCursor<Document> cursor = foodCollection.find(
                 eq(Food.NAME, regex)).iterator()){
             while(cursor.hasNext()){
                 foods.add(foodFromDocument(cursor.next()));
@@ -299,11 +375,10 @@ public class MongoDB{
 
     public Food lookUpMostEatenFoodByCategory(String category){
         openConnection();
-
         Food currentFood, targetFood = null;
-        MongoCollection<Document> dietCollection = database.getCollection(COLLECTION_FOODS);
+        MongoCollection<Document> foodCollection = database.getCollection(COLLECTION_FOODS);
 
-        try(MongoCursor<Document> cursor = dietCollection.find(
+        try(MongoCursor<Document> cursor = foodCollection.find(
                 eq(Food.CATEGORY, category)).iterator()){
             while(cursor.hasNext()){
                 currentFood = foodFromDocument(cursor.next());
@@ -332,6 +407,8 @@ public class MongoDB{
     }
 
     public boolean addEatenFood(StandardUser standardUser, EatenFood eatenFood){
+        // add id to EatenFood object
+        // update eatenTimesCount field.
         openConnection();
         MongoCollection<Document> userCollection = database.getCollection(COLLECTION_USERS);
